+++ packages/core-js/es/regexp/test.js
@@ -1,0 +1,5 @@
+require('../../modules/es.regexp.test');
+
+module.exports = function (re, string) {
+  return RegExp.prototype.test.call(re, string);
+};